requirements.txt
Flask==2.3.3
python-telegram-bot==20.3
requests==2.31.0
gunicorn==20.1.0